+++ gix-revision/tests/spec/parse/anchor/colon_symbol.rs
@@ -1,4 +1,4 @@
-use git_revision::spec;
+use gix_revision::spec;
 
 use crate::spec::parse::{parse, try_parse};