+++ gix-revision/tests/spec/parse/navigate/colon_symbol.rs
@@ -1,4 +1,4 @@
-use git_revision::spec::parse::delegate::Traversal;
+use gix_revision::spec::parse::delegate::Traversal;
 
 use crate::spec::parse::{parse, PeelToOwned as PeelTo};