+++ gix-revision/tests/spec/parse/navigate/tilde_symbol.rs
@@ -1,4 +1,4 @@
-use git_revision::{spec, spec::parse::delegate::Traversal};
+use gix_revision::{spec, spec::parse::delegate::Traversal};
 
 use crate::spec::parse::{parse, try_parse};